+++ front/src/components/calendar/MonthCalendar/style.module.css
@@ -1,0 +1,11 @@
+.calendar_table { width: 100%; height: 100%; text-align: center; border-collapse: collapse; table-layout: fixed; } 
+.calendar_table thead { width: 100%; height: 5px; vertical-align: middle; } 
+.calendar_table tbody { width: 100%; height: 95%; vertical-align: top; } 
+
+.calendar_table * {
+    user-select: none; /* supported by Chrome and Opera */
+   -webkit-user-select: none; /* Safari */
+   -khtml-user-select: none; /* Konqueror HTML */
+   -moz-user-select: none; /* Firefox */
+   -ms-user-select: none; /* Internet Explorer/Edge */
+}
