+++ front/src/store/events.js
@@ -19,7 +19,6 @@ const { selectAll } = eventsAdapter.getSelectors();
 const initialState = {
   ...eventsAdapter.getInitialState(),
   byDate: {},
-  newEventBars: [],
 };
 
 const events = createSlice({
@@ -29,23 +28,6 @@ const events = createSlice({
     updateEventBar(state) {
       const { selectAll } = eventsAdapter.getSelectors();
       state.byDate = classifyEventsByDate(selectAll(state));
-    },
-    setNewEventBars(state, { payload }) {
-      state.newEventBars = payload;
-    },
-    updateNewEventBarProperty(state, { payload }) {
-      const { key, value } = payload;
-      state.newEventBars = state.newEventBars.map(eventBar => ({
-        ...eventBar,
-        [key]: value,
-      }));
-    },
-    updateNewEventBarProperties(state, { payload }) {
-      state.newEventBars = state.newEventBars.map(eventBar =>
-        payload.reduce((info, { key, value }) => ({ ...info, [key]: value }), {
-          ...eventBar,
-        }),
-      );
     },
   },
 
@@ -81,12 +63,7 @@ const events = createSlice({
   },
 });
 
-export const {
-  updateEventBar,
-  setNewEventBars,
-  updateNewEventBarProperty,
-  updateNewEventBarProperties,
-} = events.actions;
+export const { updateEventBar } = events.actions;
 export default events.reducer;
 
 function classifyEventsByDate(events) {
